requirements_minimal.txt
flask
flask-httpauth
flask-limiter
flask-cors
werkzeug
requests
python-dotenv
cryptography
bcrypt
coloredlogs
jinja2
colorama
gunicorn
psutil
uvicorn
fastapi
loguru
httpx
openai==1.70.0
pyjwt
passlib
aiohttp==3.11.16
numpy==2.1.2
python-multipart
websockets
watchdog
